fix build

+++ src/db/main-db.cpp
@@ -131,7 +131,7 @@ MainDb::MainDb (const shared_ptr<Core> &core) : AbstractDb(*new MainDbPrivate), 
 		// `row id` is not supported by soci on Sqlite3. It's necessary to cast id to int...
 		return q->getBackend() == AbstractDb::Sqlite3
 			? static_cast<long long>(row.get<int>(0))
-			: row.get<unsigned long long>(0);
+			: static_cast<long long>(row.get<unsigned long long>(0));
 	}
 
 	long long MainDbPrivate::insertSipAddress (const string &sipAddress) {
@@ -1870,10 +1870,9 @@ MainDb::MainDb (const shared_ptr<Core> &core) : AbstractDb(*new MainDbPrivate), 
 			tm lastUpdateTime = row.get<tm>(4);
 			int capabilities = row.get<int>(5);
 			string subject = row.get<string>(6);
-			unsigned int lastNotifyId = static_cast<unsigned int>(getBackend() == Backend::Mysql
+			unsigned int lastNotifyId = (getBackend() == Backend::Mysql)
 				? row.get<unsigned int>(7, 0)
-				: row.get<int>(7, 0)
-			);
+				: static_cast<unsigned int>(row.get<int>(7, 0));
 
 			if (capabilities & static_cast<int>(ChatRoom::Capabilities::Basic)) {
 				chatRoom = core->getPrivate()->createBasicChatRoom(
